requirements.txt
flask
pyyaml
transformers
torch
numpy
scikit-learn